+++ changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2017-10-19
+### Added
+- Maintenance controller target link in config.php
+- itemlist.php view fragment
+- Mtce.php maintenance controller
+
+### Changed
+- Markdown error in readme.md
+
 ## [1.0.0] - 2017-10-19
 ### Changed
 - Readme to reflect "release" of lab 5 and beginning of lab 6